Combine all Evaluator files into one.
