+++ MetaphysicsIndustries.Solus.csproj
@@ -154,7 +154,6 @@
     <Compile Include="Macros\SubstMacro.cs" />
     <Compile Include="MathObjectHelper.cs" />
     <Compile Include="Settings.cs" />
-    <Compile Include="Evaluator.Eval.cs" />
     <Compile Include="Transformers\CleanUpTransformer.cs" />
     <Compile Include="Transformers\CloneTransformer.cs" />
     <Compile Include="Transformers\DerivativeTransformer.cs" />
@@ -168,7 +167,6 @@
     <Compile Include="SolusAction.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Evaluator.cs" />
-    <Compile Include="Evaluator.GatherVariables.cs" />
     <Compile Include="SolusParser.cs" />
     <Compile Include="SolusGrammar.cs" />
     <Compile Include="SolusEnvironment.cs" />
